fix: move vector types into numeric package
